factor out showLogs types
